Still changing things
Somethings in model i changed but not done. added tables, conclusion draft,  

Important!:  the methodology, proof sections of the body, i think i want you to write somethings for this, i think it will be alot better if you write in some proofs and i pad it.  I didn't add the time thing in i was talking about because I'm not sure how you feel about me not knowing how to derive that equation :P  im not going to write a huge thing just to have to change the whole thing directly after i finish it.